fix: Validação de usuário existente em publicação de stream.

+++ app/server/api/public/stream/publish.post.ts
@@ -1,18 +1,28 @@
 import { exec } from "child_process";
 import { JwtAdapter } from "~/src/jwt/JwtAdapter";
 import { Stream } from "~/src/streams/Stream";
+import { StreamPayload } from "~/src/streams/StreamPayload";
 import { StreamRepository } from "~/src/streams/StreamRepository";
+import { UserRepository } from "~/src/users/UserRepository";
 
 export default defineEventHandler(async (event) => {
     const body: Stream = await readBody(event);
     const jwt = JwtAdapter.builder();
+    const userRepository = new UserRepository()
+    const streamRepository = new StreamRepository();
     
     if (!jwt.verify(body.key)) {
         setResponseStatus(event, 403)
         return;
     }
 
-    const streamRepository = new StreamRepository();
+    const payload = jwt.decode(body.key) as StreamPayload
+    const user = await userRepository.findById(payload.id);
+    if (!user) {
+        setResponseStatus(event, 403)
+        return;
+    }
+
     let isInLive = await streamRepository.findByName(body.name);
     if (isInLive) {
         await streamRepository.deleteByNameAndApp(body)
